requirements.txt
paho-mqtt==1.6.1
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
xgboost==1.7.4
pyspark==3.3.2
psycopg2-binary==2.9.5
boto3==1.26.137
requests==2.28.2
schedule==1.1.0
PyYAML==6.0
joblib==1.2.0
torch==2.0.0
transformers==4.28.1